+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * st-sidebar: Allow to customize some styles from outside using css variables
 * st-table: Allow to customize sort and filter icons using an input
 * st-pagination: Allow to customize icon classes of previous and next page buttons using input "iconClasses"
+* st-pop-over: Allow to choose placement and added input 'openToLeft' to be displayed from the button left side 
 
 **Breaking changes:**